src/utils/requirements.txt
bibtexparser
pillow
PyMuPDF